add_Statistic_and_refactoring

+++ src/main/java/com/example/demo/controllers/SurveyController.java
@@ -5,7 +5,14 @@ import com.example.demo.dto.AccessData;
 import com.example.demo.dto.AccessDataResponse;
 import com.example.demo.dto.SurveyWithId;
 import java.util.UUID;
+
+import com.example.demo.dto.dtoForParseBody.AccessDataAndEmail;
+import com.example.demo.dto.dtoForParseBody.AnswerOnSurveyAndEmail;
+import com.example.demo.dto.dtoForParseBody.SurveyDataAndEmail;
+import com.example.demo.services.CheckingAvailabilityService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 //@CrossOrigin(origins = "http://localhost:3000")
@@ -14,62 +21,78 @@ public class SurveyController {
 
     private final DbService dbService;
 
+    private final CheckingAvailabilityService checkingAvailabilityService;
+
     @Autowired
-    public SurveyController(DbService dbService) {
+    public SurveyController(DbService dbService, CheckingAvailabilityService checkingAvailabilityService) {
         this.dbService = dbService;
+        this.checkingAvailabilityService = checkingAvailabilityService;
     }
 
-    @PostMapping("/user/{email}/survey")
-    public String addNewSurvey(@PathVariable("email") String email,
-                               @RequestBody String body) {
-        var id = dbService.addSurvey(email, body);
-        return id.toString();
+    @PostMapping("/survey")
+    public ResponseEntity<String> addNewSurvey(@RequestBody SurveyDataAndEmail body) {
+        if (!checkingAvailabilityService.isAvailableEmail(body.email))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        var id = dbService.addSurvey(body.email, body.surveyData);
+        return new ResponseEntity<>(id.toString(), HttpStatus.OK);
     }
 
-    @GetMapping("/user/{email}/survey/{survey_id}")
-    public String returnSurvey(@PathVariable("email") String email,
-                               @PathVariable("survey_id") UUID surveyId) {
+    @GetMapping("/survey/{survey_id}")
+    public ResponseEntity<String> returnSurvey(@PathVariable("survey_id") UUID surveyId) {
+        if (!checkingAvailabilityService.isAvailableSurvey(surveyId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         var surveyBody = dbService.getSurvey(surveyId);
-        return surveyBody.getSurvey();
+        return new ResponseEntity<>(surveyBody.getSurvey(), HttpStatus.OK);
     }
 
-    @PostMapping(value = "/user/{email}/survey/{survey_id}/access")
-    public void addSurveyAccess(@PathVariable("survey_id") UUID surveyId,
-                                @RequestBody AccessData accessData) {
-        dbService.addSurveyAccess(surveyId, accessData);
+    @PostMapping(value = "/survey/{survey_id}/access")
+    public ResponseEntity<Void> addSurveyAccess(@PathVariable("survey_id") UUID surveyId,
+                                @RequestBody AccessDataAndEmail accessData) {
+        dbService.addSurveyAccess(surveyId, accessData.accessData);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @GetMapping("/user/{email}/survey/{survey_id}/access")
-    public AccessDataResponse getSurveyAccess(@PathVariable("survey_id") UUID surveyId) {
-        return dbService.getSurveyAccess(surveyId);
+    @GetMapping("/survey/{survey_id}/access")
+    public ResponseEntity<AccessDataResponse> getSurveyAccess(@PathVariable("survey_id") UUID surveyId) {
+        return new ResponseEntity<>(dbService.getSurveyAccess(surveyId), HttpStatus.OK);
     }
 
 
-    @DeleteMapping("/user/{email}/survey/{survey_id}")
-    public void deleteSurvey(@PathVariable("survey_id") UUID surveyId) {
+    @DeleteMapping("/survey/{survey_id}")
+    public ResponseEntity<Void> deleteSurvey(@PathVariable("survey_id") UUID surveyId) {
+        if (!checkingAvailabilityService.isAvailableSurvey(surveyId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         dbService.deleteSurvey(surveyId);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @PatchMapping("/user/{email}/survey/{survey_id}")
-    public void updateSurvey(@PathVariable("survey_id") UUID surveyId,
-                             @RequestBody String body) {
-        dbService.updateSurvey(surveyId, body);
+    @PatchMapping("/survey/{survey_id}")
+    public ResponseEntity<Void> updateSurvey(@PathVariable("survey_id") UUID surveyId,
+                             @RequestBody SurveyDataAndEmail body) {
+        if (!checkingAvailabilityService.isAvailableSurvey(surveyId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        dbService.updateSurvey(surveyId, body.surveyData);
+        return new ResponseEntity<>(HttpStatus.OK);
     }
 
-    @GetMapping("/user/{email}/surveys")
-    public SurveyWithId[] returnSurveys(@PathVariable("email") String email) {
+    @GetMapping("/surveys")
+    public ResponseEntity<SurveyWithId[]> returnSurveys(@RequestParam String email) {
+        if (!checkingAvailabilityService.isAvailableEmail(email))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         var surveys = dbService.getSurveys(email);
         var surveysWithId = surveys.stream()
                 .map(val -> new SurveyWithId(val.getId(), val.getSurvey()))
                 .toArray(SurveyWithId[]::new);
-        return surveysWithId;
+        return new ResponseEntity<>(surveysWithId, HttpStatus.OK);
     }
 
-    @PostMapping("/user/{email}/survey/{survey_id}/answer")
-    public String getAnswerOnSurvey(@PathVariable("email") String email,
-                                    @PathVariable("survey_id") UUID surveyId,
-                                    @RequestBody String body) {
-        var id = dbService.addResponseOnSurvey(email, surveyId, body);
-        return id.toString();
+
+    @PostMapping("/survey/{survey_id}/answer")
+    public ResponseEntity<String> getAnswerOnSurvey(@PathVariable("survey_id") UUID surveyId,
+                                    @RequestBody AnswerOnSurveyAndEmail body) {
+        if (!checkingAvailabilityService.isAvailableSurvey(surveyId))
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        var id = dbService.addResponseOnSurvey(body.email, surveyId, body.answer);
+        return new ResponseEntity<>(id.toString(), HttpStatus.OK);
     }
 }
